разделяет моковые данные и константы
