+++ src/main.js
@@ -6,3 +6,4 @@ const siteMainElement = document.querySelector('.trip-events');
 const waypointModel = new WaypointModel();
 const presenter = new TripPresenter({ headerContainer: siteFiltersElement, mainContainer: siteMainElement, waypointModel });
 presenter.init();
+document.body.addEventListener('click', (evt) => console.log(evt.target))
