+++ src/mock/data.js
@@ -1,0 +1,14 @@
+const TYPE = ['taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'];
+const CITIES = ['Amsterdam', 'Chamonix', 'Geneva'];
+const DESCRIPTION = [
+  'Located on the ocean with white sandy beaches and crystal clear water',
+  'Ancient castle with towers and a moat around it',
+  'Mountain resort with beautiful views and skiing opportunities',
+  'Huge park with diverse flora and fauna',
+  'Historic city center with many landmarks'
+];
+const PHOTO = 'https://loremflickr.com/248/152?random=';
+const OFFER_TITLE = ['Add luggage', 'Switch to comfort class', 'Add meal', 'Choose seats', 'Travel by train', 'Order Uber'];
+
+
+export { TYPE, CITIES, DESCRIPTION, PHOTO, OFFER_TITLE };
