+++ src/view/no-event.js
@@ -1,5 +1,5 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import { TextNoEvent } from '../mock/constants.js';
+import { TextNoEvent } from '../utils/constants.js';
 
 function createNoEventTemplate() {
   return (`<p class="trip-events__msg">${TextNoEvent.EVERYTHING}</p>`);
